stinger_net needs to include header directly from stinger_alg source dir.
Because headers are build targets now, CMake needs to know which targets depend on which headers. stinger_alg depends on stinger_net, so all the headers are guaranteed to be copied before stinger_alg is run. But stinger_alg_client needs to include streaming_algorithm.h, which belongs to stinger_alg, so now we have a circular dependency. I solved it by explicity including streaming_algorithm.h from the source directory.

In the future, stinger_alg should be refactored to not depend on stinger_net, so this will become a one-way dependency.